Database::Money removed int #931
git-svn-id: file:///data/svn/enum/server/trunk@17396 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/cli_admin/handle_adminclientselection_args.h
@@ -1676,7 +1676,7 @@ public:
                 ::value<Checked::string>()->notifier(save_string(params.valid_to))
                 , "price valid to datetime")
             ("operation_price", boost::program_options
-                ::value<Checked::ulong>()->notifier(save_arg<long>(params.operation_price))
+                ::value<Checked::ulong>()->notifier(save_arg<long long>(params.operation_price))
                 , "operation price")
             ("period", boost::program_options
                 ::value<Checked::ulong>()->notifier(save_ulong(params.period))
