+++ src/cli_admin/registrar_params.h
@@ -336,7 +336,7 @@ struct PriceAddArgs
 {
     optional_string valid_from;
     optional_string valid_to;
-    long operation_price;
+    long long operation_price;
     optional_ulong period;
     optional_string zone_fqdn;
     optional_id zone_id;
@@ -351,7 +351,7 @@ struct PriceAddArgs
     PriceAddArgs(
             const optional_string& _valid_from
             , const optional_string& _valid_to
-            , const long _operation_price
+            , const long long _operation_price
             , const optional_ulong& _period
             , const optional_string& _zone_fqdn
             , const optional_id& _zone_id
